A not so pretty hack for Windows OF 0.9.8.
Hacks and work arounds and bug fixes are implemented for Windows OF 0.9.8.
All the modified codes are enclosed with preprocessor #if OF_VER_09X ~ #endif, which is defined in "ofxShadowMap.h".
Thus, if it were error-clean on Mac/Linux before modifications, it should work fine even after modifications.

Added "extMaterial" .cpp&.hpp and "shaders/phong" .frag&.vert to "src" directory, which are fundamentally OF 0.10.X codes, but inherits and overrides the current(which is older) version of class ofMaterial to be compatible with the current version of class ofGLProgrammableRenderer.

Also fixed some bugs that render wrong shadow texture sampling and mapping, yet not sure whether Windows or OF being not 0.10.X is the culprit for the mess.